+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Tokens: Add `font`, `font-family` and `line-height` css tokens.
 - Feature: Utilities: Automagically generate utility classes from tokens.
 - Fix: `Button`: Fix `hydrogen` button style with `icon` fit.
+- Feature: `SlideModal`: Add new component.
 
 ## [11.7.1] - 2022-07-07